entrega 4 pronta

+++ structures.h
@@ -5,9 +5,9 @@ typedef enum {FALSE,TRUE} BOOL;
 
 typedef struct
 {
-    int wall;
-    int group;
-    int sz;
+    short int wall;
+    short int group;
+    short int sz;
 } cell;
 
 typedef enum {A1 = 1, A2, A3, A4, A5, A6} mode;
